notemind_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
psycopg2-binary==2.9.9
# Web Framework
fastapi
uvicorn

# LangChain & GigaChat
langgraph
langchain-gigachat

# Other
python-dotenv
salute-speech
openrouteservice